requirements.txt
langchain
langchain-groq
python-dotenv
streamlit
streamlit-chat
